port2 Phase 2: Global variable and macro cleanup
- Replace var macro with extern, rewrite ex_extern.c with explicit definitions
- Replace typedef short bool with <stdbool.h>
- Remove ignore()/ignorf() lint macros, update copy() macro comment
- Widen short types to int for screen coords, state vars, struct members
- Remove #define var from vi_mem.c

+++ port2/ex_get.c
@@ -19,7 +19,7 @@
  * we have different flavors of routines which do/don't return such.
  */
 static	bool junkbs;
-short	lastc = '\n';
+int	lastc = '\n';
 
 ignchar()
 {
