+++ port2/ex_temp.c
@@ -25,7 +25,7 @@
 char	tfname[40];
 char	rfname[40];
 int	havetmp;
-short	tfile = -1;
+int	tfile = -1;
 short	rfile = -1;
 
 fileinit()
